fix: update build to work with new cpy version
